+++ keycloak-wsfed/src/main/java/io/github/chosomeister/keycloak/protocol/wsfed/WSFedLoginProtocol.java
@@ -223,7 +223,7 @@ public class WSFedLoginProtocol implements LoginProtocol {
                     .setAction(WSFedConstants.WSFED_SIGNIN_ACTION)
                     .setDestination(clientSession.getRedirectUri())
                     .setContext(context)
-                    .setTokenExpiration(io.github.chosomeister.keycloak.protocol.wsfed.builders.WsFedSAMLAssertionTypeAbstractBuilder.tokenLifespan(realm, client))
+                    .setTokenExpiration(io.github.chosomeister.keycloak.protocol.wsfed.builders.WsFedSAMLAssertionTypeAbstractBuilder.tokenLifespan(realm, client, userSession))
                     .setRequestIssuer(clientSession.getClient().getClientId());
 
             configureTokenSecurity(session, realm, client, ctx.getBuilder());